+++ R/cbd_kenpom_ratings.R
@@ -2,8 +2,6 @@
 #'
 #' Pulls year-end KenPom ratings and adjusted efficiencies from 2001-Present.
 #'
-#' Requires KenPom authorization. You can authorize with
-#' `cbd_kenpom_authorization`.
 #'
 #' @param ... OPTIONAL. To load the complete data set, pass no arguments through
 #'   to the function. Else, you can filter on  `team`, `conf`, `year`, or any